+++ src/main/groovy/io/xh/hoist/cluster/BaseClusterResult.groovy
@@ -1,0 +1,4 @@
+package io.xh.hoist.cluster
+
+interface BaseClusterResult {}
+