+++ src/main/groovy/io/xh/hoist/cluster/JsonClusterResult.groovy
@@ -1,0 +1,8 @@
+package io.xh.hoist.cluster
+
+class JsonClusterResult implements BaseClusterResult {
+    String value
+    boolean valueIsVoid
+    String exception
+    Integer exceptionStatusCode
+}